- 새로운 디렉토리 구조 반영
