--- src/pages/About.tsx
@@ -1,7 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router';
-
-import Test from '@components/test';
-
-export const Route = createFileRoute('/about')({
-  component: Test,
-});
