+++ src/pages/__root.tsx
@@ -8,9 +8,6 @@ export const Route = createRootRoute({
         <Link to="/" className="[&.active]:font-bold">
           Home
         </Link>{' '}
-        <Link to="/about" className="[&.active]:font-bold">
-          About
-        </Link>{' '}
         <Link to="/dashboard" className="[&.active]:font-bold">
           dashboard
         </Link>
