+++ src/pages/dashboard/-components/Card/index.tsx
@@ -1,0 +1,5 @@
+import { Container } from './styled';
+
+export default function Card() {
+  return <Container>Card</Container>;
+}
