+++ src/pages/dashboard/-components/Card/styled.tsx
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material/styles';
+
+export const Container = styled('div')(({ theme }) => ({
+  display: 'flex',
+  flexDirection: 'column',
+  backgroundColor: theme.baseColor.base200,
+  color: theme.baseColor.base100,
+}));
