+++ src/pages/dashboard/index.tsx
@@ -1,9 +1,12 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import Layout from './-components/layout';
+import Card from './-components/Card';
+import Layout from './-layout';
+import { Container } from './-styled';
 
 export const Route = createFileRoute('/dashboard/')({
   component: () => (
+    // NOTE 레이아웃 컴포넌트를 여기에서 랩핑할건지 밑에서 랩핑할건지 결정필요.
     <Layout>
       <Dashboard />
     </Layout>
@@ -11,5 +14,10 @@ export const Route = createFileRoute('/dashboard/')({
 });
 
 function Dashboard() {
-  return <div>dashboard mainPage</div>;
+  return (
+    <Container>
+      dashboard mainPage
+      <Card />
+    </Container>
+  );
 }
